interval parameter added to simulation

+++ scripts/aavs2/start_aavs2_sensitivity_simul.sh
@@ -5,6 +5,10 @@ if [[ -n "$1" && "$1" != "-" ]]; then
    channel=$1
 fi
 
+interval=86400
+if [[ -n "$2" && "$2" != "-" ]]; then
+   interval=$2
+fi
 
 path=`pwd`
 dir_path=`dirname $path`
@@ -19,10 +23,14 @@ export PATH=~/github/station_beam/:/opt/anaconda2/bin:$PATH
 # echo "nohup calc_eda_sensitvity_all.sh $ux_start 86400 - \"--use_beam_fits --station_name=EDA --size=512 --trcv_type=trcv_eda2\" $mwa_coarse_channel - 300 400 400 > out 2>&1 &"
 # nohup calc_eda_sensitvity_all.sh $ux_start 86400 - "--use_beam_fits --station_name=EDA --size=512 --trcv_type=trcv_eda2" $mwa_coarse_channel - 300 400 400 > out 2>&1 &
 
-echo "cp ~/aavs-calibration/config/aavs2/antenna_locations.txt antenna_locations_aavs2.txt"
-cp ~/aavs-calibration/config/aavs2/antenna_locations.txt antenna_locations_aavs2.txt
+if [[ ! -s antenna_locations_aavs2.txt ]]; then
+   echo "cp ~/aavs-calibration/config/aavs2/antenna_locations.txt antenna_locations_aavs2.txt"
+   cp ~/aavs-calibration/config/aavs2/antenna_locations.txt antenna_locations_aavs2.txt
+else
+   echo "File antenna_locations_aavs2.txt already exist -> copy skipped"
+fi
 
-echo "nohup calc_eda_sensitvity_all.sh $ux_start 86400 - "--use_beam_fits --station_name=SKALA4 --size=512 --trcv_type=trcv_aavs2_vs_za_deg --antenna_locations=antenna_locations_aavs2.txt" $mwa_coarse_channel - 300 400 400 > out 2>&1 &"
-nohup calc_eda_sensitvity_all.sh $ux_start 86400 - "--use_beam_fits --station_name=SKALA4 --size=512 --trcv_type=trcv_aavs2_vs_za_deg --antenna_locations=antenna_locations_aavs2.txt" $mwa_coarse_channel - 300 400 400 > out 2>&1 &
+echo "nohup calc_eda_sensitvity_all.sh $ux_start ${interval} - "--use_beam_fits --station_name=SKALA4 --size=512 --trcv_type=trcv_aavs2_vs_za_deg --antenna_locations=antenna_locations_aavs2.txt" $mwa_coarse_channel - 300 400 400 > out 2>&1 &"
+nohup calc_eda_sensitvity_all.sh $ux_start ${interval} - "--use_beam_fits --station_name=SKALA4 --size=512 --trcv_type=trcv_aavs2_vs_za_deg --antenna_locations=antenna_locations_aavs2.txt" $mwa_coarse_channel - 300 400 400 > out 2>&1 &
 
 # ~/github/station_beam/tools/merge_sensitivity_loop.sh
